refactor: change from snapshot config to more general integration config
