+++ dev-client/__tests__/snapshot/LoginScreen-test.tsx
@@ -15,7 +15,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import {render} from '@testing/snapshot/utils';
+import {render} from '@testing/integration/utils';
 
 import {LoginScreen} from 'terraso-mobile-client/screens/LoginScreen';
